Tried Fixing the Flex Boxes, unsuccessful again.
I changed the classes and structure of the code a bit but it's still not complete. I have asked for help in two separate coding help discords and have yet to receive a good response. I have been googling for the past three days trying to figure this out. Hopefully, I can resolve this issue by the end of the week, or this Sunday, the 26th.