80% resolved flexbox issue, adjusted text size
found a stack overflow thread that had the "solution" to my issue. It just had a different variation that I liked. It is more feasible than the vision I had, but it will do. Also noticed text size was way too big for the mobile versions, so I made them smaller and will increase the size for laptops/tablets with media queries. That's right, mobile-first design 😎.