Added about.html, fixed footer, finished home page
Finished the home page for now. Don't think I will have anything else to style or add, at least for now. Finally began working on other pages. Originally was going to work on the signup page but decided to just make the about.html page and work on that. Didn't do much on it, just added the footer and navbar, both of which are fully functional.