New stylesheet, began adding text for about page
Decided 1 stylesheet was going to be way too messy and confusing for a website of this size, so I made a styles folder with a stylesheet for the two pages I'm working on. I'll add more when I need to. Spent a lot of time breaking my head trying to figure out how to style this page. Was very difficult but I have a very basic idea of what I will do. Hopefully, I'll get more inspiration in the coming days as I work on it. Let's see what happens.