blog like button, gotta fix footer
I made the blog like button work (i need to save it in MongoDB so people can only like once and the amount is the same for every user). Realized the footer is broken on most if not every page, so i have to redesign it. preferably from scratch so i can fully understand it.